chore: move all tests into package tests
