+++ .coveragerc
@@ -8,3 +8,4 @@ omit =
     *admin.py
     */static/*
     */templates/*
+    **/tests/*
