+++ platform_plugin_aspects/tests/test_course_published.py
