+++ platform_plugin_aspects/tests/test_signals.py
