+++ test_settings.py
@@ -1,3 +1,4 @@
+
 """
 These settings are here to use during tests, because django requires them.
 
@@ -5,48 +6,48 @@ In a real-world use case, apps in this project are installed into other
 Django applications, so these settings will not be used.
 """
 
-from os.path import abspath, dirname, join
-
-
-def root(*args):
-    """
-    Get the absolute path of the given path relative to the project root.
-    """
-    return join(abspath(dirname(__file__)), *args)
-
-DEBUG = True
-
 DATABASES = {
-    'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': 'default.db',
-        'USER': '',
-        'PASSWORD': '',
-        'HOST': '',
-        'PORT': '',
+    "default": {
+        "ENGINE": "django.db.backends.sqlite3",
+        "NAME": "default.db",
+        "USER": "",
+        "PASSWORD": "",
+        "HOST": "",
+        "PORT": "",
+    },
+    "read_replica": {
+        "ENGINE": "django.db.backends.sqlite3",
+        "NAME": "read_replica.db",
+        "USER": "",
+        "PASSWORD": "",
+        "HOST": "",
+        "PORT": "",
+    },
+}
+
+
+INSTALLED_APPS = (
+    "platform_plugin_aspects",
+)
+
+EVENT_SINK_CLICKHOUSE_MODEL_CONFIG = {
+    "user_profile": {
+        "module": "common.djangoapps.student.models",
+        "model": "UserProfile",
+    },
+    "course_overviews": {
+        "module": "openedx.core.djangoapps.content.course_overviews.models",
+        "model": "CourseOverview",
     }
 }
 
-INSTALLED_APPS = (
-    'django.contrib.admin',
-    'django.contrib.auth',
-    'django.contrib.contenttypes',
-    'django.contrib.messages',
-    'django.contrib.sessions',
-    'platform_plugin_aspects',
-)
+EVENT_SINK_CLICKHOUSE_COURSE_OVERVIEWS_ENABLED = True
 
-LOCALE_PATHS = [
-    root('platform_plugin_aspects', 'conf', 'locale'),
-]
+FEATURES = {
+    'CUSTOM_COURSES_EDX': True,
+}
 
-SECRET_KEY = 'insecure-secret-key'
-
-MIDDLEWARE = (
-    'django.contrib.auth.middleware.AuthenticationMiddleware',
-    'django.contrib.messages.middleware.MessageMiddleware',
-    'django.contrib.sessions.middleware.SessionMiddleware',
-)
+DEBUG = True
 
 TEMPLATES = [{
     'BACKEND': 'django.template.backends.django.DjangoTemplates',
