+++ tox.ini
@@ -30,13 +30,13 @@ ignore = D101,D200,D203,D212,D215,D404,D405,D406,D407,D408,D409,D410,D411,D412,D
 match-dir = (?!migrations)
 
 [pytest]
-DJANGO_SETTINGS_MODULE = test_utils.test_settings
+DJANGO_SETTINGS_MODULE = test_settings
 addopts = --cov platform_plugin_aspects --cov-report term-missing --cov-report xml --log-level=INFO
 norecursedirs = .* docs requirements site-packages
 
 [testenv]
 setenv:
-    DJANGO_SETTINGS_MODULE = test_utils.test_settings
+    DJANGO_SETTINGS_MODULE = test_settings
 deps =
     django32: Django>=3.2,<4.0
     django40: Django>=4.0,<4.1
@@ -47,7 +47,7 @@ commands =
 
 [testenv:docs]
 setenv =
-    DJANGO_SETTINGS_MODULE = test_utils.test_settings
+    DJANGO_SETTINGS_MODULE = test_settings
     PYTHONPATH = {toxinidir}
     # Adding the option here instead of as a default in the docs Makefile because that Makefile is generated by shpinx.
     SPHINXOPTS = -W
@@ -73,9 +73,7 @@ allowlist_externals =
 deps =
     -r{toxinidir}/requirements/quality.txt
 commands =
-    touch tests/__init__.py
     pylint platform_plugin_aspects tests test_utils manage.py setup.py
-    rm tests/__init__.py
     pycodestyle platform_plugin_aspects tests manage.py setup.py
     pydocstyle platform_plugin_aspects tests manage.py setup.py
     isort --check-only --diff tests test_utils platform_plugin_aspects manage.py setup.py
@@ -83,7 +81,7 @@ commands =
 
 [testenv:pii_check]
 setenv =
-    DJANGO_SETTINGS_MODULE = test_utils.test_settings
+    DJANGO_SETTINGS_MODULE = test_settings
 deps =
     -r{toxinidir}/requirements/test.txt
 commands =
